requirements.txt
h5py==3.10.0
hdf5plugin==4.3.0
matplotlib==3.8.1
netCDF4==1.6.5
numpy==1.26.1
pandas==2.1.2
PyYAML==6.0.1
scipy==1.11.3
setuptools==68.0.0
xarray==2023.10.1
yamlloader==1.3.2
